refactor: clean up demo frontend
